Unify int-to-string conversion.
The codebase was full of repeated patterns for:

1. Representing a nonnegative number as a string.
2. Representing a small nonnegative number as a digit.
3. Functions for representing numbers as string, mostly unused.

This replaces all of them with a single new inline function.  In the case
where the output was meant to be one digit, there were cases where actually
the number could overflow, resulting in weird characters.  In particular,
these could cause crashes on Windows machines with more than 10 CPU cores.
You'd get a filename with ':' in it, which doesn't work.  Such cases are
replaced with zero-padded three-digit strings, to keep sort orders sane.