Deal with random numbers outside target interval.
The zufall() ("coincidence," "random chance") function gets a random double
in the half-open interval [0, 1), and scales it to fall in the half-open
interval [min, max).  It asserts that the result is in that interval.  But
rounding can throw an occasional spanner in the works, and produce e.g. a
value that's equal to max.  When that happens, just throw the dice again.