Fix array overrun.
Looks like MYOptimization was added as an afterthought but never fully
implemented.  An array was created to cover all optimization classes
except this one, and then the entry for MYOptimization was initialized
beyond the end.

Since all other code appeared to ignore the entry anyway, I simply
removed the poisonous initialization.